fix: 오류 수정

+++ lib/controllers/game_session_controller.dart
@@ -155,7 +155,7 @@ class GameSessionController extends ChangeNotifier {
   void completeSession({required int rawScore}) {
     _isCompleted = true;
     _timer?.cancel();
-    _finalScore = (rawScore - hintPenalty).clamp(0, 100);
+    _finalScore = (rawScore - hintPenalty).clamp(0, 100).toInt();
     notifyListeners();
   }
